FEAT: 为本地渲染战绩提供更多特性

+++ src/main/kotlin/action/commandProcess/eternalReturn/entity/profile/EternalReturnProfilePlayer.kt
@@ -9,5 +9,5 @@ data class EternalReturnProfilePlayer(
     val lastPlayedSeasonId: Int,
     val name: String,
     val syncedAt: Long,
-    val userNum: Int
+    val userNum: Long,
 )
